requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
python-dotenv
pypdf2
tiktoken